# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample pocket server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "version": "1.0.6",
        "title": "Swagger Pocket",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:4000",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/forgotuser": {
            "post": {
              "tags": [
                "user"
              ],
              "summary": "send email",
              "description": "send email by forgot password",
              "operationId": "create",
              "consumes": [
                "application/json"
              ],
              "produces": [
                "application/json",
                "application/xml"
              ],
              "parameters": [
                {
                  "in": "body",
                  "name": "body",
                  "description": "send email",
                  "required": true,
                  "schema": {
                    "$ref": "#/definitions/forgotuser"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "successful operation",
                  "schema": {
                    "$ref": "#/definitions/forgotuser"
                  }
                },
                "400": {
                  "description": "Invalid email"
                }
              }
            }
          },
          "/verifyuserotp": {
            "post": {
              "tags": [
                "user"
              ],
              "summary": "verify otp",
              "description": "verify otp",
              "operationId": "create",
              "consumes": [
                "application/json"
              ],
              "produces": [
                "application/json",
                "application/xml"
              ],
              "parameters": [
                {
                  "in": "body",
                  "name": "body",
                  "description": "send email",
                  "required": true,
                  "schema": {
                    "$ref": "#/definitions/verifyuserotp"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "successful operation",
                  "schema": {
                    "$ref": "#/definitions/verifyuserotp"
                  }
                },
                "400": {
                  "description": "Invalid email"
                }
              }
            }
          },
...
          "/sendmail-otp":{
            "post": {
                "tags": [
                  "user"
                ],
                "summary": "send otp by email",
                "description": "send email by otp",
                "operationId": "create",
                "consumes": [
                  "application/json"
                ],
                "produces": [
                  "application/json",
                  "application/xml"
                ],
                "parameters": [
                  {
                    "in": "body",
                    "name": "body",
                    "description": "send email",
                    "required": true,
                    "schema": {
                      "$ref": "#/definitions/sendmail-otp"
                    }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "successful operation",
                    "schema": {
                      "$ref": "#/definitions/sendmail-otp"
                    }
                  },
                  "400": {
                    "description": "Invalied otp or email"
                  }
                }
              }
        },
        "/phonenumber-otp":{
            "post": {
                "tags": [
                  "user"
                ],
                "summary": "send otp by phonenumber",
                "description": "send otp by phonenumber",
                "operationId": "create",
                "consumes": [
                  "application/json"
                ],
                "produces": [
                  "application/json",
                  "application/xml"
                ],
                "parameters": [
                  {
                    "in": "body",
                    "name": "body",
                    "description": "send email",
                    "required": true,
                    "schema": {
                      "$ref": "#/definitions/phonenumber-otp"
                    }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "successful operation",
                    "schema": {
                      "$ref": "#/definitions/phonenumber-otp"
                    }
                  },
                  "400": {
                    "description": "Invalied otp or email"
                  }
                }
              }
        },
        "/forgotusermobile":{
            "post": {
                "tags": [
                  "user"
                ],
                "summary": "send otp by phonenumber",
                "description": "send otp by phonenumber",
                "operationId": "create",
                "consumes": [
                  "application/json"
                ],
                "produces": [
                  "application/json",
                  "application/xml"
                ],
                "parameters": [
                  {
                    "in": "body",
                    "name": "body",
                    "description": "send email",
                    "required": true,
                    "schema": {
                      "$ref": "#/definitions/forgotusermobile"
                    }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "successful operation",
                    "schema": {
                      "$ref": "#/definitions/forgotusermobile"
                    }
                  },
                  "400": {
                    "description": "Invalied otp or email"
                  }
                }
              }
        },
        "/verifyuserupdatepassword":{
            "post": {
                "tags": [
                  "user"
                ],
                "summary": "send otp by phonenumber",
                "description": "send otp by phonenumber",
                "operationId": "create",
                "consumes": [
                  "application/json"
                ],
                "produces": [
                  "application/json",
                  "application/xml"
                ],
                "parameters": [
                  {
                    "in": "body",
                    "name": "body",
                    "description": "send email",
                    "required": true,
                    "schema": {
                      "$ref": "#/definitions/verifyuserupdatepassword"
                    }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "successful operation",
                    "schema": {
                      "$ref": "#/definitions/verifyuserupdatepassword"
                    }
                  },
                  "400": {
                    "description": "Invalied otp or email"
                  }
                }
              }
        },
...
      },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        },
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
            "flow": "implicit",
            "scopes": {
                "read:pets": "read your pets",
                "write:pets": "modify pets in your account"
            }
        }
    },
    "definitions": {
...
          "forgotuser": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string"
              }
            },
            "xml": {
              "name": "ngo"
            }
          },
          "verifyuserotp": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string"
              },
              "otp":{
                "type":"integer"
              }
             
            },
            "xml": {
              "name": "ngo"
            }
          },
...
          "sendmail-otp": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string"
              },
              "otp":{
                "type":"integer"
              }
             
            },
            "xml": {
              "name": "user"
            }
          },
          "phonenumber-otp": {
            "type": "object",
            "properties": {
              "phonenumber": {
                "type": "string"
              },
              "otp":{
                "type":"integer"
              }
             
            },
            "xml": {
              "name": "user"
            }
          },
          "forgotusermobile": {
            "type": "object",
            "properties": {
              "phonenumber": {
                "type": "string"
              }
            },
            "xml": {
              "name": "user"
            }
          },
          "verifyuserupdatepassword": {
            "type": "object",
            "properties": {
              "phonenumber": {
                "type": "string"
              },
              "otp":{
                "type":"integer"
              }
             
            },
            "xml": {
              "name": "ngo"
            }
          },
...
    }
  }